+++ TailorSwift/Classes/CollectionAdditions.swift
@@ -10,32 +10,26 @@ import Foundation
 
 extension Collection {
 
-    // swiftlint:disable:next variable_name
-    internal func _expectEnd<C: Collection>(of s: C, is i: C.Index) {
-        _debugPrecondition(
-            i == s.endIndex,
-            "invalid Collection: count differed in successive traversals")
+    public func element(at index: Self.Index) -> Self.Iterator.Element? {
+        if endIndex > index && startIndex <= index {
+            return self[index]
+        } else {
+            return nil
+        }
     }
-
-    /// Whether or not the collection is __not__ empty.
-    /// returns: A `Bool` indicating whether or not the collection is __not__ empty.
-    public var isNotEmpty: Bool {
-        return !isEmpty
-    }
-
-    public func mapDistinct<T: Equatable>(_ transform: (Iterator.Element) throws -> T
-        ) rethrows -> [T] {
-
+    
+    public func uniqueMap<T: Equatable>(_ transform: (Iterator.Element) throws -> T) rethrows -> [T] {
+        
         let count: Int = numericCast(self.count)
         if isEmpty {
             return []
         }
-
+        
         var result = ContiguousArray<T>()
         result.reserveCapacity(count)
-        // swiftlint:disable:next identifier_name
+        
         var i = self.startIndex
-
+        
         for _ in 0..<count {
             let transformed = try transform(self[i])
             if !result.contains(where: { $0 == transformed }) {
@@ -43,8 +37,6 @@ extension Collection {
             }
             formIndex(after: &i)
         }
-
-        _expectEnd(of: self, is: i)
         return Array(result)
     }