+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2019-07-04
+### Added
+- Support for Swift Package Manager using Xcode 11.0 beta.
+
+### Changed
+- `mapDistinct(_:)` renamed to `uniqueMap(_:)` to better fit with naming of `compactMap(_:)` and `flatMap(_:)`.
+
 ## [1.0.0] - 2019-01-22
 ### Added
 - Added `isNotEmpty` property to `String` and `Collection`.